pyproject.toml
[project]
dependencies = [
    "msal >= 1.34.0",
    "openpyxl >= 3.1.5",
    "packaging >= 25.0",
    "pandas >= 2.3.3",
    "requests >= 2.32.5",
    "requests-cache >= 1.2.1",
    "rich >= 14.2.0",
    # TODO: Pin to the next release.
    "sdmx1 @ git+https://github.com/khaeru/sdmx.git@f57d7f5be8249027bf63d5198a54351019b045a2",
]
name = "sdmx-explorer"
version = "0.1.0"
description = "An interactive SDMX explorer and downloader."
authors = [{ name = "Ben Frankel", email = "ben.frankel7@gmail.com" }]
readme = "README.md"

[project.scripts]
explore = "sdmx_explorer.explore:main"
download = "sdmx_explorer.download:main"
